requirements.txt
langchain
langchainhub
python-dotenv
fastapi
uvicorn
gunicorn
loguru
chromadb
sentence_transformers
transformers